requirements.txt
alembic==0.9.9
algopy==0.5.7
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
Babel==2.5.3
bkcharts==0.2
blinker==1.4
bokeh==0.12.16
cffi==1.11.5
chardet==2.3.0
click==6.7
cloud-init==17.2
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
coverage==4.5.1
cryptography==2.1.4
cssselect==1.0.3
cycler==0.10.0
decorator==4.2.1
Flask==1.0.2
Flask-BabelEx==0.9.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Principal==0.4.0
Flask-Restless==0.17.0
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
hibagent==1.0.1
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
ipython==6.2.1
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10
jsonpatch==1.10
jsonpointer==1.9
kiwisolver==1.0.1
language-selector==0.1
lxml==4.1.1
Mako==1.0.3
MarkupSafe==1.0
matplotlib==2.2.2
mimerender==0.6.0
more-itertools==4.1.0
networkx==2.1
numdifftools==0.9.20
numpy==1.14.3
oauthlib==1.0.3
packaging==17.1
pandas==0.23.0
pandas-highcharts==0.5.2
parsel==1.4.0
parso==0.1.1
passlib==1.7.1
patsy==0.5.0
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.1.0
pluggy==0.6.0
prettytable==0.7.2
prompt-toolkit==1.0.15
psycopg2==2.6.1
ptyprocess==0.5.2
py==1.5.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
pycurl==7.43.0
PyDispatcher==2.0.5
pyflux==0.4.15
Pygments==2.2.0
pygobject==3.20.0
PyJWT==1.3.0
pyOpenSSL==17.5.0
pyparsing==2.2.0
pyserial==3.0.1
pytest==3.5.0
pytest-splinter==1.8.5
python-apt==1.1.0b1+ubuntu0.16.4.1
python-dateutil==2.7.3
python-debian==0.1.27
python-editor==1.0.3
python-mimeparse==1.6.0
python-systemd==231
pytz==2018.4
PyWavelets==0.5.2
PyYAML==3.12
queuelib==1.4.2
requests==2.9.1
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.1.0
Scrapy==1.5.0
selenium==3.11.0
service-identity==17.0.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
speaklater==1.3
splinter==0.7.7
SQLAlchemy==1.2.7
ssh-import-id==5.5
tornado==5.0.2
traitlets==4.3.2
Twisted==17.9.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.13.1
virtualenv==15.0.1
vulture==0.26
w3lib==1.19.0
wcwidth==0.1.7
Werkzeug==0.14.1
WTForms==2.1
zope.interface==4.4.3
